transactions | db-lost | order api external api

+++ server/src/persistance/mongoDb/GeneralCollection.ts
@@ -1,48 +1,76 @@
-import {  Model } from "mongoose";
+import { Session } from "inspector";
+import { ITransaction } from "../Icollection";
+
+var mongoose = require('mongoose');
 export const createCollection = (TModel,extractT) =>
     class {
+
         static async drop (){
-           await TModel.collection.drop();
-           
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
+            await TModel.collection.drop();
         }
-        static async find(opt={}){
-            // const keys = Object.keys(opt).forEach((key) => {
-            //     if(opt[key].$in)
-                    
-            // })
-            const modelsDb = await TModel.find(opt);
-            return modelsDb.map(extractT);
-        };
-        static async findByIds(ids=[]){
-            const modelsDb = await TModel.find({ _id: { $in: ids}});
+
+        static async find(filter, opt:{session?:any}={}){
+            if(mongoose.connection.readyState === 0){
+                throw new Error('connection lost');
+            }
+            const modelsDb = await TModel.find(filter,null,
+                opt.session ? {session: opt.session} :{}) ;
             return modelsDb.map(extractT);
         };
 
-        static async findOne(opt={}){
-            const res = await TModel.findOne(opt);
+        static async findByIds(ids=[],opt:{session?:any}={}){
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
+            const modelsDb = await TModel.find({ _id: { $in: ids}},null,
+                opt.session ? {session: opt.session} :{}) ;
+            return modelsDb.map(extractT);
+        };
+
+        static async findOne(filter, opt:{session?:any}={}){
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
+            const res = await TModel.findOne(filter,null,
+                opt.session ? {session: opt.session} :{});
             return res ? extractT(res) : null;
         };
 
-        static async findById(id){
-            const res = await TModel.findById(id);
+        static async findById(id, opt:{session?:any}={}){
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
+            const res = await TModel.findById(id,null,
+                opt.session ? {session: opt.session} :{}) ;
             return res ? extractT(res) : null;
         };
 
-        static async updateOne(model){
-            const modelDb = await TModel.findById(model.id || model._id);
+        static async updateOne(model, opt: {session?:any}={}){
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
+            let modelDb = await TModel.findById(model.id || model._id,null, 
+                opt.session ? {session:opt.session} :{}) ;
             updateMongoFields(modelDb,model);
             
             const res = await modelDb.save();
             return res? extractT(res) : null
         }
 
-        static async delete(opt){
-            await TModel.deleteOne(opt);
+        static async delete(filter, opt: {session?:any}={}){
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
+            await TModel.deleteMany(filter,
+                opt.session ? {session:opt.session} :{}) ;
             return null;
         }
 
-        static async insert(model){ 
+        static async insert(model,opt: {session?:any}={}){ 
+            await TModel.createCollection();
+            if(mongoose.connection.readyState === 0)
+                throw new Error('connection lost');
             let newTmodel = new TModel({}); 
+            if(opt.session)
+                newTmodel.$session(opt.session);
+
             updateMongoFields(newTmodel,model);
             
             newTmodel = await newTmodel.save();
@@ -60,3 +88,30 @@ export const createCollection = (TModel,extractT) =>
             }
         });
     }
+
+
+export class Transaction implements ITransaction {
+        _session;
+
+        async startTransaction (){
+            this._session = await mongoose.connection.startSession();
+            await this._session.startTransaction();
+            return ({session: this._session});
+        }
+
+        async commitTransaction (){
+            await this._session.commitTransaction();
+            await this._session.endSession();
+        }
+
+        async abortTransaction (){
+            await this._session.abortTransaction();
+            await this._session.endSession();
+        }
+
+        session(){
+            return this._session;
+        }
+    }
+
+
